chore/SV-208: CustomMarker -> CustomMarkerComponent로 이름 변경
